test: only force test suite exit on WOA

+++ spec/index.js
@@ -4,6 +4,8 @@ const fs = require('node:fs');
 const path = require('node:path');
 const v8 = require('node:v8');
 
+const { FAILURE_STATUS_KEY } = require('../script/spec-runner');
+
 // We want to terminate on errors, not throw up a dialog
 process.on('uncaughtException', (err) => {
   console.error('Unhandled exception in main spec runner:', err);
@@ -163,8 +165,12 @@ app.whenReady().then(async () => {
   const cb = () => {
     // Ensure the callback is called after runner is defined
     process.nextTick(() => {
-      console.log(`failures: ${runner.failures}`);
-      process.kill(process.pid);
+      if (process.env.ELECTRON_FORCE_TEST_SUITE_EXIT) {
+        console.log(`${FAILURE_STATUS_KEY}: ${runner.failures}`);
+        process.kill(process.pid);
+      } else {
+        app.exit(runner.failures);
+      }
     });
   };
